refactor(literature): split the check harness, and make the measurement scripts reproducible
Two groups of review points from mrj4001 on the dev-only harnesses.

STRUCTURE. literatureSearch.check.js was 1064 lines. It is now a thin entry
point — `npm run check:literature` is unchanged — over per-concern files:

  .check.harness.js   ROOT/OUT, loadEnvLocal, cleanOutDir, compileLiterature
                      (execSync wrapped so a compile failure names the sources
                      and the out dir), requireCompiled
  .check.fixtures.js  the shared fixtures
  .check.mode1/2/3.js one file per mode
  .check.rowCounts.js the count-map assertions
  .check.dialects.js  per-database syntax and seed quoting
  .check.exports.js   the export contract

Stays plain JS with node: prefixed core imports rather than becoming
TypeScript: it is deliberately runnable with no build step.

REPRODUCIBILITY. These scripts are the measurement artifacts, and none of
them recorded enough to answer "exactly what produced this number?".

  recall.js   artifact reshaped to { metadata, reviews }, where metadata
              carries ISO timestamp, git SHA, model, region, record cap, rank
              depth and the benchmark path; rows now carry the strategy and
              criteria too. Throws on duplicate gold PMIDs, which would have
              silently inflated a denominator. Screen recall no longer
              collapses three outcomes into one number — kept, excluded and
              no-verdict are reported separately against the same denominator.
              DEEP is unchanged at 1000, but the labels now say "within the
              first 1000" instead of implying a full rank.

  screen.js   writes .litscreen-run.json with the same metadata block, on the
              stop path too. Every stage prints its realised prevalence rather
              than a hardcoded claim: stage A was described as "~50% eligible"
              and is actually 36-60%, and stage B is 6%, not "realistic". The
              stop rule now fires on the pooled threshold OR any single review
              below it, so one domain failure cannot hide behind two successes.
              Gold sampling sorts by PMID before slicing, so PubMed's return
              order is no longer the sampling order. Adds one assertion that a
              verdict actually reaches the "AI suggested" column of
              recordSheets() — the model-to-Covidence path was untested.
              Comments now state what the cost figure excludes, and why the
              distractor include rate is not a precision score.

  question.js artifact carries the generated strategy and criteria, not just
              the score, since model sampling makes a re-run non-identical.
              buildStrategy's positional arguments are named at the call site;
              its signature is unchanged because production code calls it too.

  pure.check  tightened five comment blocks that restated themselves. The
              central claim — that this check reads no .env.local, and that
              this is itself an assertion — is untouched.

Thresholds, model ids and benchmark fixtures are unchanged; none of these
scripts was run as part of this commit.
