+++ .gitignore
@@ -43,3 +43,4 @@ node_modules
 .litrecall/
 .litrecall-run.json
 .litquestion-run.json
+.litscreen-run.json
