+++ controllers/literatureSearch.check.js
@@ -30,1035 +30,56 @@
  * papers, so the count drifts; asserting `hits === 122` would rot within weeks. What is
  * asserted instead is RELATIONAL (dropping an AND-ed block can only widen the search) and
  * BEHAVIOURAL (these four fixed historical papers hit or miss for a derivable reason).
+ *
+ * AND IT STAYS PLAIN JS, deliberately. `node controllers/literatureSearch.check.js` runs with no
+ * build step, no ts-node, no transpile of its own — the only compile in the whole harness is the
+ * one it performs on the modules UNDER TEST. Rewritten in TypeScript, the safety net would need the
+ * toolchain to be working before it could tell you the toolchain had broken, and the first thing
+ * anyone does with a check that will not start is stop running it.
+ *
+ * THIS FILE IS THE ENTRY POINT AND THE RUNNING ORDER; the assertions live in the siblings named
+ * below, the fixtures in .fixtures.js and the plumbing in .harness.js. The order is not cosmetic:
+ * Mode 1 counts the fixture and hands the yield to Mode 2, the row counts are what the exports
+ * appendix is checked against, and the Embase run is asserted in the dialects part and exported in
+ * the exports part.
  */
-const fs = require('fs')
-const path = require('path')
-const assert = require('assert')
-const { execSync } = require('child_process')
+const {
+    cleanOutDir, compileLiterature, loadEnvLocal, requireCompiled,
+} = require('./literatureSearch.check.harness')
+const mode1 = require('./literatureSearch.check.mode1')
+const mode2 = require('./literatureSearch.check.mode2')
+const mode3 = require('./literatureSearch.check.mode3')
+const rowCounts = require('./literatureSearch.check.rowCounts')
+const dialects = require('./literatureSearch.check.dialects')
+const exportsCheck = require('./literatureSearch.check.exports')
 
-const ROOT = path.resolve(__dirname, '..')
-// Compile INSIDE the worktree, not os.tmpdir(). The controller imports
-// @aws-sdk/client-bedrock-runtime at module scope, and node resolves node_modules by walking
-// UP from the required file — from /var/folders/... it never reaches the repo and the require
-// dies with MODULE_NOT_FOUND. (The check still makes no model call; the import just has to
-// resolve.) .litcheck/ is gitignored.
-const OUT = path.join(ROOT, '.litcheck')
-
-// Embase (Ovid) is asserted in two halves: the pure/dialect facts where the dialects are checked, and
-// the EXPORT in the exports section, where the renderers are loaded. This carries the run between them.
-let embaseRun = null
-fs.rmSync(OUT, { recursive: true, force: true })
-
-// A laptop keeps RECITER_PUBMED_API_URL in .env.local, which is gitignored — so on any runner that
-// checks the repo out there is no such file, and reading it unconditionally crashes with ENOENT
-// before a single assertion runs. A missing file is therefore not an error: the same names arrive as
-// real environment variables instead. An UNSET url still is an error, and the first count says so
-// loudly. Where the file does exist it still WINS, or a stale shell export would quietly point a
-// local run at a different PubMed than the one the developer is reading in .env.local.
-const envFile = path.join(ROOT, '.env.local')
-if (fs.existsSync(envFile)) {
-    for (const line of fs.readFileSync(envFile, 'utf8').split('\n')) {
-        const m = line.match(/^([A-Z_]+)\s*=\s*(.*)$/)
-        if (m) process.env[m[1]] = m[2].trim().replace(/^["']|["']$/g, '')
-    }
-}
-
-try {
-    execSync(
-        `npx tsc controllers/literatureSearch.controller.ts controllers/literatureExport.ts ` +
-        `controllers/literatureDocx.ts --outDir ${OUT} ` +
-        `--module commonjs --target es2020 --esModuleInterop --skipLibCheck --allowJs`,
-        { cwd: ROOT, stdio: 'inherit' },
-    )
-} catch (err) {
-    throw new Error(`Failed to compile literature modules for the check: ${err.message}`)
-}
-const lit = require(path.join(OUT, 'controllers/literatureSearch.controller.js'))
-
-// A strategy shaped exactly as the model is prompted to emit one: PRESS-style, each concept
-// split into its MeSH line and its free-text line, limits held separately. Two things here are
-// load-bearing. Fully tagged: PubMed's automatic term mapping rewrites untagged terms, which
-// would make the count irreproducible. Separate lines: a checkbox has to have a line to sit on.
-const strategy = () => ({
-    db: 'pubmed',
-    concepts: [
-        {
-            label: 'Probiotics / microbiome',
-            lines: [
-                { terms: '"Gastrointestinal Microbiome"[MeSH] OR "Probiotics"[MeSH]', on: true },
-                { terms: 'probiotic*[tiab] OR synbiotic*[tiab]', on: true },
-            ],
-        },
-        {
-            label: 'Depression',
-            lines: [
-                { terms: '"Depression"[MeSH] OR "Depressive Disorder"[MeSH]', on: true },
-                { terms: 'depress*[tiab]', on: true },
-            ],
-        },
-    ],
-    limits: '(2021:2026[dp]) AND (Randomized Controlled Trial[pt])',
-})
-
-// Untick every line of one concept — what a librarian does to the "Adults" block that was
-// silently killing their seeds.
-const untickConcept = (s, ci) => ({
-    ...s,
-    concepts: s.concepts.map((c, i) =>
-        i === ci ? { ...c, lines: c.lines.map(l => ({ ...l, on: false })) } : c),
-})
+cleanOutDir()
+loadEnvLocal()
+compileLiterature([
+    'controllers/literatureSearch.controller.ts',
+    'controllers/literatureExport.ts',
+    'controllers/literatureDocx.ts',
+])
+const lit = requireCompiled('controllers/literatureSearch.controller.js')
 
 ;(async () => {
-    // ---- Pure: assembleQuery + numberStrategy. No network. --------------------------------
-    const s = strategy()
-
-    const q = lit.assembleQuery(s)
-    assert.ok(q.includes(' AND '), 'concepts are AND-ed')
-    assert.ok(q.includes(' OR '), 'the lines within a concept are OR-ed')
-    assert.ok(q.endsWith(s.limits), 'limits are appended, not folded into a concept')
-
-    // THE ONE THAT MATTERS. A concept with nothing ticked must DROP OUT OF THE AND entirely.
-    // If it instead emitted `(...) AND ()`, PubMed would either reject the query or — far worse —
-    // silently reinterpret it, and the librarian would be handed a count for a search they never
-    // ran. Every checkbox on the page rests on this being right.
-    const dropped = lit.assembleQuery(untickConcept(s, 1))
-    assert.ok(!/\(\s*\)/.test(dropped), 'an unticked concept must never emit an empty ()')
-    assert.ok(!dropped.includes('depress'), 'the unticked concept is gone from the query')
-    assert.strictEqual(
-        dropped,
-        `("Gastrointestinal Microbiome"[MeSH] OR "Probiotics"[MeSH] OR probiotic*[tiab] OR synbiotic*[tiab]) AND ${s.limits}`,
-        'one live concept: no stray AND, no stray parens',
-    )
-
-    // Nothing ticked at all is not "0 hits" — it is "there is no strategy". Empty string, so
-    // runStrategy refuses to count it rather than reporting a confident, meaningless zero.
-    assert.strictEqual(lit.assembleQuery(untickConcept(untickConcept(s, 0), 1)), '', 'no live concepts -> no query')
-
-    // PRESS numbering is DERIVED from the selection, so it can never describe a line that was
-    // not searched: 1,2 = probiotics lines, 3 = 1 OR 2, 4,5 = depression, 6 = 4 OR 5,
-    // 7 = 3 AND 6, 8 = 7 AND limits.
-    const { rows, conceptLines } = lit.numberStrategy(s)
-    assert.deepStrictEqual(conceptLines, [[1, 2], [4, 5]], 'each concept occupies its own numbered lines')
-    assert.deepStrictEqual(
-        rows.filter(r => r.kind === 'combine').map(r => `${r.n}: ${r.text}`),
-        ['3: 1 OR 2', '6: 4 OR 5', '7: 3 AND 6', `8: 7 AND ${s.limits}`],
-        'the combination lines recompute from the ticked lines',
-    )
-    // Untick one line and the numbering closes up behind it — the numbers ARE the current
-    // selection. The free-text probiotics line goes; a single-line concept needs no OR row, so
-    // probiotics now resolves straight to line 1.
-    const oneOff = {
-        ...s,
-        concepts: s.concepts.map((c, i) => i === 0 ? { ...c, lines: [c.lines[0], { ...c.lines[1], on: false }] } : c),
-    }
-    assert.deepStrictEqual(
-        lit.numberStrategy(oneOff).rows.filter(r => r.kind === 'combine').map(r => `${r.n}: ${r.text}`),
-        ['4: 2 OR 3', '5: 1 AND 4', `6: 5 AND ${s.limits}`],
-        'unticking a line renumbers, and a single-line concept emits no OR row',
-    )
-    // ...and THAT is why a rowCounts map cannot outlive the selection it was counted for. The map is
-    // keyed by line number, and a line number does not name a stable query: here, number 4 means
-    // "Depression"[MeSH] before the untick and "2 OR 3" after it. Read the old map at the new numbers
-    // and every row shows some other row's count -- silently, and including the final row, the one
-    // that IS the search. The UI must therefore DROP the map on any edit (dropRowCounts), not grey it
-    // out. This asserts the hazard rather than the fix, because the fix lives in React and this file
-    // cannot render: if a number ever stops being re-keyed by an untick, this goes red and the
-    // invalidation can be revisited.
-    const numberedQuery = st => new Map(lit.numberStrategy(st).rows.filter(r => r.n !== null).map(r => [r.n, r.query]))
-    const before = numberedQuery(s)
-    const after = numberedQuery(oneOff)
-    const collision = [...after].filter(([n, q]) => before.has(n) && before.get(n) !== q)
-    assert.ok(
-        collision.length > 0,
-        'unticking a line RE-KEYS the rows: a count map from the previous selection must be dropped, never re-read',
-    )
-    assert.notStrictEqual(
-        before.get(Math.max(...before.keys())),
-        after.get(Math.max(...after.keys())),
-        'even the final row -- the whole search -- changes its number, so a stale map mislabels the yield line itself',
-    )
-
-    // ---- Live PubMed. ---------------------------------------------------------------------
-    const hits = await lit.countPubmed(q)
-    assert.ok(Number.isFinite(hits) && hits > 0, `expected a positive yield, got ${hits}`)
-    console.log(`yield: ${hits}`)
-
-    // A seed is an IDENTIFIER WITH A KIND, never a bare PMID — a Scopus-only record has no PMID at
-    // all. parseSeeds is what the route and the browser both use, so it is what is checked.
-    const seedList = lit.parseSeeds([
-        '37314797', // Nikolova 2023  JAMA Psychiatry    -> in the set
-        '35654766', // Schaub   2022  Transl Psychiatry  -> in the set
-        '34875345', // Tian     2022  Brain Behav Immun  -> in the set
-        '27793434', // Sarkar   2016  Trends Neurosci    -> MISS
-    ])
-    assert.strictEqual(seedList.length, 4, 'four PMIDs parse out of the seeds box')
-    assert.ok(seedList.every(x => x.kind === 'pmid'), 'digits are PMIDs')
-
-    const seeds = await lit.validateSeeds(s, seedList)
-
-    for (const x of seeds) {
-        const why = x.retrieved
-            ? ''
-            : `  <- fails: ${[...x.failingConcepts.map(i => s.concepts[i].label), ...(x.failsLimits ? ['LIMITS'] : [])].join(', ')}`
-        console.log(`  ${x.retrieved ? 'HIT ' : 'MISS'}  ${x.kind.toUpperCase()} ${x.id}${why}`)
-    }
-
-    assert.strictEqual(seeds.filter(x => x.retrieved).length, 3, 'three seeds must be retrieved')
-
-    // Sarkar 2016 CLEARS the probiotics block (it is MeSH-indexed under Probiotics) but FAILS
-    // the depression block: PubMed indexes it under Emotions, and its abstract never says
-    // "depression". A model asked to GUESS the reason would most likely have blamed the
-    // 2021-2026 date range. Deriving the reason rather than guessing it is what makes the miss
-    // trustworthy, so it is what this check defends.
-    const miss = seeds.find(x => x.id === '27793434')
-    assert.ok(miss && !miss.retrieved, 'Sarkar 2016 must miss this strategy')
-    assert.deepStrictEqual(miss.failingConcepts, [1], 'the miss is attributed to the Depression block (index 1)')
-
-    // AND IT ALSO FAILS THE LIMITS — it is a 2016 review, and we asked for 2021-2026 RCTs. Both
-    // causes are reported, independently, because reporting only the block sends a librarian off
-    // to widen a search that STILL cannot return the paper. This exact hole shipped once: the
-    // suggested "Emotions" line was verified against the block alone, so the page promised
-    // "retrieves Sarkar (2016)", charged +531 records to screen, and did not retrieve it.
-    assert.strictEqual(miss.failsLimits, true, 'Sarkar 2016 also fails the 2021-2026 RCT limits')
-
-    // So the honest end-to-end check: widening the Depression block does NOT bring it back while
-    // those limits stand. This is the assertion that would have caught the false promise.
-    const widenedDepression = {
-        ...s,
-        concepts: s.concepts.map((c, i) => i === 1
-            ? { ...c, lines: [...c.lines, { terms: '"Emotions"[MeSH] OR mood[tiab]', on: true }] }
-            : c),
-    }
-    assert.strictEqual(
-        await lit.countPubmed(`27793434[uid] AND (${lit.assembleQuery(widenedDepression)})`), 0,
-        'widening the block alone must NOT retrieve a seed the limits exclude',
-    )
-    // Drop the limits as well and it finally comes back — proving the limits were the second,
-    // independent cause rather than a red herring.
-    assert.strictEqual(
-        await lit.countPubmed(`27793434[uid] AND (${lit.assembleQuery({ ...widenedDepression, limits: '' })})`), 1,
-        'widening the block AND dropping the limits retrieves it',
-    )
-
-    // The toggle arithmetic the whole feature rests on: dropping an AND-ed block can only widen
-    // the search. This is what a librarian sees when they untick the block that was killing their
-    // seeds, and it is why the checkboxes are safe to hand over.
-    const widened = await lit.countPubmed(lit.assembleQuery(untickConcept(s, 1)))
-    assert.ok(widened > hits, `unticking a concept must widen the search (${hits} -> ${widened})`)
-    console.log(`untick the Depression block: ${hits} -> ${widened} records`)
-
-    // THE MONOTONICITY INVARIANT, and the assertion that would have caught a real bug.
-    //
-    // OR-ing terms INTO a block can only ADD records. `widened >= base` is arithmetic. When it is
-    // violated, the maths is not wrong — THE COUNT IS. A throttled esearch comes back as a
-    // well-formed 0 rather than an error (unkeyed NCBI is 3 req/s and a build fires a dozen
-    // counts), and on 2026-07-13 that produced a suggested widening priced at "+-5,714 records":
-    // countPubmed returned 0 for a query that reliably counts 64,604.
-    //
-    // suggestFixes now refuses to publish a price that violates this, and countPubmed retries a
-    // zero once. This is the check that keeps both honest.
-    const widenedBlockHits = await lit.countPubmed(lit.assembleQuery(widenedDepression))
-    assert.ok(
-        widenedBlockHits >= hits,
-        `widening a block cannot SHRINK the search — the count is untrustworthy (${hits} -> ${widenedBlockHits})`,
-    )
-    console.log(`widen the Depression block:  ${hits} -> ${widenedBlockHits} records (+${widenedBlockHits - hits})`)
-
-    // =======================================================================================
-    // MODE 2 — "Issue review". Records, not counts. Still no LLM: everything below is either
-    // pure or a PubMed fetch, and the two things it defends are the two that silently produce
-    // a page that looks perfect and is worthless.
-
-    // ---- Pure: the design chip. -------------------------------------------------------------
-    // Derived from PubMed's OWN publication types, never from the model's read of the abstract —
-    // the chip is a claim about the evidence hierarchy, and PubMed's indexers already made it.
-    assert.strictEqual(lit.designOf(['Randomized Controlled Trial', 'Journal Article']), 'RCT')
-    assert.strictEqual(lit.designOf(['Journal Article', 'Review', "Research Support, Non-U.S. Gov't"]), 'Review')
-    assert.strictEqual(lit.designOf(['Observational Study']), 'Observational')
-    assert.strictEqual(lit.designOf(['Journal Article']), 'Other', 'a bare Journal Article is "Other", not a guess')
-    assert.strictEqual(lit.designOf([]), 'Other')
-    // The types OVERLAP, so the order is the evidence hierarchy and the strongest true label wins.
-    // A meta-analysis is ALSO tagged Systematic Review and Review; showing it as "Review" would
-    // demote the strongest evidence on the page.
-    assert.strictEqual(
-        lit.designOf(['Meta-Analysis', 'Systematic Review', 'Review', 'Journal Article']), 'Meta-analysis',
-        'meta-analysis outranks the Review tags PubMed also puts on it',
-    )
-    // Qualified variants are real ("Randomized Controlled Trial, Veterinary"), and the tags arrive
-    // in PubMed's own casing.
-    assert.strictEqual(lit.designOf(['randomized controlled trial, veterinary']), 'RCT')
-
-    // ---- Pure: the structured-abstract join. ------------------------------------------------
-    // PubMed abstracts arrive in SEGMENTS. Keeping the labels is what lets a model tell a stated
-    // RESULT from a stated OBJECTIVE — the exact distinction the synthesis prompt polices.
-    assert.strictEqual(
-        lit.joinAbstract([
-            { abstractTextLabel: 'OBJECTIVE', abstractText: 'To test X.' },
-            { abstractTextLabel: 'METHODS', abstractText: 'An RCT.' },
-        ]),
-        'OBJECTIVE: To test X.\nMETHODS: An RCT.',
-        'the segments are joined with their labels kept',
-    )
-    // An unlabelled abstract is one segment and must NOT acquire a label out of thin air.
-    assert.strictEqual(lit.joinAbstract([{ abstractText: 'One flat paragraph.' }]), 'One flat paragraph.')
-    assert.strictEqual(lit.joinAbstract(undefined), '', 'a record with no abstract joins to empty, never to a crash')
-    // THE TRAP, pinned. The field is `abstractText` — camelCase — while its siblings on `article`
-    // are lowercase. The obvious guess parses cleanly, throws nothing, and yields an empty string:
-    // 50 records screened on their titles alone, which reads plausible and is worthless.
-    assert.strictEqual(lit.joinAbstract([{ abstracttext: 'wrong casing' }]), '', 'the lowercase guess yields NOTHING')
-
-    // ---- Pure: the self-cancelling filter. ---------------------------------------------------
-    // Lines within a block are OR-ed. So "Humans"[MeSH] OR-ed into a study-design block is
-    // satisfied by every human paper ever indexed, and the design filter silently becomes a no-op.
-    // The model did exactly this in a live run; 31 of 50 retrieved records came back reviews.
-    {
-        const bad = {
-            db: 'pubmed',
-            limits: '(2021:2026[dp])',
-            concepts: [
-                { label: 'Probiotics', lines: [{ terms: 'Probiotics[MeSH] OR probiotic[tiab]', on: true }] },
-                { label: 'Study design', lines: [{ terms: 'Randomized Controlled Trial[pt] OR Humans[MeSH] OR placebo[tiab]', on: true }] },
-            ],
-        }
-        const fixed = lit.hoistFilters(bad)
-        const design = fixed.concepts.find(c => c.label === 'Study design')
-        assert.ok(!/Humans/i.test(design.lines[0].terms), 'Humans is pulled OUT of the design block')
-        assert.ok(/Randomized Controlled Trial\[pt\]/.test(design.lines[0].terms), 'the real design terms survive')
-        const humans = fixed.concepts.find(c => c.label === 'Humans')
-        assert.ok(humans, 'Humans gets its OWN block, so that it is AND-ed rather than OR-ed')
-        // The whole point: it must now AND, i.e. actually restrict.
-        assert.ok(lit.assembleQuery(fixed).includes(') AND (Humans[MeSH])'), 'and the AND is real in the emitted query')
-
-        // A filter already alone in its own block is left exactly as it is — no churn.
-        const good = {
-            db: 'pubmed', limits: '',
-            concepts: [{ label: 'Humans', lines: [{ terms: 'Humans[MeSH]', on: true }] }],
-        }
-        assert.deepStrictEqual(lit.hoistFilters(good), good, 'a correct strategy passes through untouched')
-    }
-
-    // ---- Live iCite: a MISSING percentile must never render as a real one. -------------------
-    // Number(null) === 0, and 0 is finite. So a finiteness check on the COERCED value cannot tell
-    // "iCite has not scored this yet" from "bottom of its field", and the first cut of this shipped
-    // 14 of 50 records — every recent trial in the set — as a confident "NIH 0th pct".
-    // 37314797 (2023) has a real percentile; 41667154 (2026) is too new to have one.
-    {
-        const scored = await lit.withCitationMetrics([
-            { pmid: '37314797', title: '', journal: '', year: '2023', authors: '', design: 'RCT', abstract: '', mesh: [] },
-            { pmid: '41667154', title: '', journal: '', year: '2026', authors: '', design: 'RCT', abstract: '', mesh: [] },
-        ])
-        const [old_, recent] = scored
-        assert.strictEqual(typeof old_.nihPercentile, 'number', 'a scored paper carries its percentile')
-        assert.ok(old_.nihPercentile > 50, 'and it is the real value, not a placeholder')
-        assert.strictEqual(recent.nihPercentile, undefined,
-            'an UNSCORED paper carries NO percentile — not 0, which would read as "bottom of its field"')
-        // The records themselves must survive regardless: this is a garnish, never the deliverable.
-        assert.strictEqual(scored.length, 2, 'iCite never drops a record')
-    }
-
-    // ---- Pure: PubMed markup. It is NOT just italics, and it is not only in titles. -----------
-    // Rendered raw, this put "Psychobiotic <i>Lactobacillus plantarum</i> JYLP-326" on the screen
-    // in a live run. In an ABSTRACT the same tags never reach the DOM — they just get billed into
-    // the context window, 50 abstracts at a time.
-    assert.strictEqual(
-        lit.plainText('Psychobiotic <i>Lactobacillus plantarum</i> JYLP-326 relieves anxiety'),
-        'Psychobiotic Lactobacillus plantarum JYLP-326 relieves anxiety',
-        'italics are stripped, not printed',
-    )
-    assert.strictEqual(lit.plainText('CO<sub>2</sub> and <sup>13</sup>C and <b>bold</b>'), 'CO2 and 13C and bold',
-        'sub/sup/bold too — the strip is generic, not an allowlist of the tags we happened to meet')
-    assert.strictEqual(lit.plainText('Bacteroides &amp; Bifidobacterium'), 'Bacteroides & Bifidobacterium',
-        'named entities are decoded')
-    assert.strictEqual(lit.plainText('SMD &#x2212;0.44'), 'SMD −0.44',
-        'numeric entities decode — an effect size must not reach the model as "&#x2212;0.44"')
-    // &amp; is decoded LAST, or "&amp;lt;" double-decodes into a real "<" and re-opens the hole.
-    assert.strictEqual(lit.plainText('a &amp;lt; b'), 'a &lt; b', 'no double-decoding into live markup')
-    assert.strictEqual(lit.plainText(undefined), '', 'a missing title is empty, never a crash')
-    // The markup strip runs on abstracts too — this is the token bug, not the render bug.
-    assert.strictEqual(
-        lit.joinAbstract([{ abstractTextLabel: 'RESULTS', abstractText: '<i>L. plantarum</i> reduced HAMD (p&lt;0.05).' }]),
-        'RESULTS: L. plantarum reduced HAMD (p<0.05).',
-        'abstracts are cleaned as well as titles',
-    )
-
-    // ---- Live PubMed: the record parse. -----------------------------------------------------
-    const recs = await lit.fetchRecords('37314797[uid] OR 35654766[uid] OR 34875345[uid] OR 27793434[uid]', 4)
-    assert.strictEqual(recs.length, 4, 'four PMIDs, four records')
-
-    // THE ASSERTION THIS WHOLE SECTION EXISTS FOR. If the abstract path ever drifts, this fails
-    // here rather than in a synthesis that quietly summarizes nothing but titles.
-    for (const r of recs) {
-        assert.ok(r.abstract.length > 200, `PMID ${r.pmid} came back with no abstract — check the camelCase path`)
-        assert.ok(['RCT', 'Meta-analysis', 'Review', 'Observational', 'Other'].includes(r.design))
-        console.log(`  ${r.pmid}  ${r.design.padEnd(13)} ${r.authors} — ${r.journal} ${r.year} (${r.abstract.length} chars)`)
-    }
-
-    const nikolova = recs.find(r => r.pmid === '37314797')
-    assert.strictEqual(nikolova.design, 'RCT', 'PubMed tags Nikolova 2023 a Randomized Controlled Trial')
-    assert.strictEqual(nikolova.authors, 'Nikolova et al.', 'first author + et al., because there are four of them')
-    assert.strictEqual(nikolova.year, '2023')
-    assert.strictEqual(nikolova.journal, 'JAMA Psychiatry', 'the ISO abbreviation, not the lowercased MEDLINE title')
-    assert.ok(nikolova.mesh.includes('Humans'), 'the MeSH descriptors come through')
-    // It is a JAMA structured abstract: eight labelled segments. The labels must survive the join.
-    assert.ok(nikolova.abstract.startsWith('IMPORTANCE: '), 'the first segment keeps its label')
-    assert.ok(nikolova.abstract.includes('\nRESULTS: '), 'every segment keeps its label, on its own line')
-
-    // Schaub 2022 is a real unlabelled abstract — the join must not invent a label for it.
-    const schaub = recs.find(r => r.pmid === '35654766')
-    assert.ok(!/^[A-Z][A-Z ,]+:/.test(schaub.abstract), 'an unlabelled abstract stays unlabelled')
-
-    const sarkar = recs.find(r => r.pmid === '27793434')
-    assert.strictEqual(sarkar.design, 'Review', 'Sarkar 2016 is tagged Review, not RCT')
-
-    // ---- Live PubMed: the cap, and the count-before-fetch rule. ------------------------------
-    // The cap is a property of the mode, not a setting, and it is enforced HERE rather than
-    // trusted to the retrieval tool's retmax: it bounds what enters the context window, so it
-    // cannot depend on the far side of an HTTP call honouring a field.
-    const capped = await lit.fetchRecords(lit.assembleQuery(s), 5)
-    assert.strictEqual(capped.length, 5, 'the cap is enforced on our side of the wire')
-
-    // THE THREE BANDS. Retrieval is always "the top 50 of N", so the only question is whether the
-    // slice is honest: <= 50 is no slice at all, <= 200 is a defensible one, above that the 50 is a
-    // thin slice of the yield and the librarian gets priced narrowings instead of a silent top-50.
-    //
-    // WHICH BAND THIS FIXTURE LANDS IN IS PUBMED'S BUSINESS — it indexes new papers every night —
-    // so what is asserted is the RULE, not the number. Asserting `hits < 200` here would rot.
-    const review = await lit.runReview(s, 'relevance')
-    assert.strictEqual(
-        !!review.needsNarrowing, review.hits > lit.NARROW_ABOVE,
-        'the narrowing panel is offered exactly when the top 50 would be a thin slice',
-    )
-    if (review.needsNarrowing) {
-        assert.strictEqual(review.records.length, 0, 'a gated search retrieves nothing until it is asked to')
-    } else {
-        assert.strictEqual(review.records.length, Math.min(review.hits, lit.RECORD_CAP), 'up to 50 records, never more')
-    }
-    assert.strictEqual(review.seeds.length, 0, 'Mode 2 has no known-item seeds, but keeps the shape')
-    console.log(`issue review: ${review.hits} hits -> ${review.records.length} records fetched`)
-
-    // A BROAD search is NOT AN ERROR, and it is NOT REFUSED. The old behaviour — a hard refusal
-    // above 2,000 hits — was never ours to impose: it was the retrieval tool's fetch limit, and the
-    // tool now takes a retmax. What is left is an honesty problem, not a technical one, so the
-    // broad search comes back with the strategy, the count, and the PRICE of each way out of it.
-    const broadStrategy = { ...s, concepts: [s.concepts[0]], limits: '' }
-    const broad = await lit.runReview(broadStrategy, 'relevance')
-    assert.ok(broad.hits > lit.NARROW_ABOVE, `expected a huge yield, got ${broad.hits}`)
-    assert.strictEqual(broad.needsNarrowing, true, 'above the band the top 50 is a thin slice, and it says so')
-    assert.strictEqual(broad.records.length, 0, 'and it retrieves nothing until the librarian has seen the price')
-    assert.ok(broad.query.length > 0, 'the strategy still comes back — it is what the librarian narrows')
-    assert.ok(broad.narrowings.length > 0, 'and it comes back with narrowings, priced')
-    console.log(`narrowing:    ${broad.hits} records match — the top 50 is a thin slice. Priced ways out:`)
-    for (const n of broad.narrowings) {
-        // EVERY published narrowing REALLY NARROWS, and the number is COUNTED, never estimated.
-        assert.ok(Number.isFinite(n.hits), `${n.label} must carry a counted yield`)
-        assert.ok(n.hits < broad.hits, `${n.label} must actually reduce the count (${broad.hits} -> ${n.hits})`)
-        assert.ok(n.terms.includes('['), `${n.label} must be fully tagged — an untagged block is irreproducible`)
-        assert.ok(!/\bAND\b|\(/.test(n.terms), `${n.label} is a BLOCK, not a query — the client ANDs it in`)
-        console.log(`  [ ] ${n.label.padEnd(24)} ${broad.hits} -> ${n.hits}   ${n.why}`)
-    }
-
-    // THE ESCAPE HATCH. `proceed` skips the gate at ANY count and retrieves the top 50 anyway. A
-    // librarian may know exactly what they are doing, and a tool that refuses to run is worse than
-    // one that warns.
-    //
-    // This is also the assertion that the deleted refusal was safe to delete: the retrieval tool
-    // takes a retmax and returns the top 50 of a query in the hundreds of thousands. If that ever
-    // stops being true, it fails HERE and not in front of a librarian.
-    const anyway = await lit.runReview(broadStrategy, 'relevance', true)
-    assert.ok(!anyway.needsNarrowing, 'proceed walks through the gate')
-    assert.strictEqual(anyway.records.length, lit.RECORD_CAP, `the top 50 of ${anyway.hits} come back regardless`)
-    console.log(`proceed:      ${anyway.hits} hits -> ${anyway.records.length} records (the top 50, retrieved anyway)`)
-
-    // ---- Live PubMed: suggestNarrowings — the mirror of the suggested widening. ---------------
-    // Same contract as suggestFixes, pointing the other way: a candidate is PRICED against live
-    // PubMed, and one whose number we cannot stand behind is DROPPED rather than published.
-    //
-    // SKIPPED WHEN THE QUERY ALREADY HAS IT. The fixture is already limited to 2021-2026 RCTs, so
-    // both the RCT block and the 5-year window are already in the assembled query — offering either
-    // would price it at a delta of zero, which is noise dressed up as advice.
-    const onFixture = await lit.suggestNarrowings(s, hits)
-    const labels = onFixture.map(n => n.label)
-    assert.ok(!labels.includes('Randomized trials only'), 'the RCT limit is already in this query')
-    assert.ok(!labels.includes('Last 5 years'), 'a date limit is already set — the dropdown owns that decision')
-    for (const n of onFixture) {
-        assert.ok(n.hits < hits, `${n.label} must actually reduce the count (${hits} -> ${n.hits})`)
-    }
-
-    // A NARROWING THAT DOES NOT REDUCE THE COUNT IS DROPPED — the assertion this section exists for,
-    // and the exact mirror of suggestFixes refusing to publish a widening that does not retrieve the
-    // seed. Useless advice with a number on it is worse than no advice: it is a checkbox that costs
-    // a click and buys nothing, next to two that are real.
-    //
-    // This strategy is ALREADY restricted to controlled clinical trials — but the words "Randomized
-    // Controlled Trial[pt]" do not appear in it, so the candidate survives the textual skip. It is
-    // offered, it is priced against live PubMed, and the price comes back IDENTICAL: every record in
-    // the base already satisfies (RCT[pt] OR CCT[pt]), so the block excludes nothing. Set theory,
-    // not a coincidence of this month's index — which is why this assertion cannot rot.
-    const cct = { ...s, limits: '(2021:2026[dp]) AND (Controlled Clinical Trial[pt])' }
-    const cctHits = await lit.countPubmed(lit.assembleQuery(cct))
-    assert.ok(cctHits > 0, `the controlled-trial fixture must find something, got ${cctHits}`)
-    const onCct = await lit.suggestNarrowings(cct, cctHits)
-    assert.ok(
-        !onCct.some(n => n.label === 'Randomized trials only'),
-        'a narrowing that leaves the count unchanged is DROPPED, never shown as a no-op with a price on it',
-    )
-    console.log(`dropped:      "Randomized trials only" prices at ${cctHits} -> ${cctHits} on a trials-only query, so it is not offered`)
-
-    // =======================================================================================
-    // MODE 3 — "Clinical question". The evidence hierarchy IS the product, so it is the thing
-    // most worth pinning. Still no LLM: the ranking is derived from PubMed's publication types,
-    // which is exactly why it can be tested at all.
-
-    // ---- Pure: the tiers. -------------------------------------------------------------------
-    // THE INVERSION THIS FIXED, PINNED SO IT CANNOT COME BACK. Before Mode 3, `systematic review`
-    // and `review` were ONE bucket tested AFTER RCT — so a systematic review sorted BELOW an RCT,
-    // and a narrative review sorted EQUAL to a systematic one. Harmless when the design was just a
-    // chip. Fatal when the ORDER IS THE ANSWER.
-    const sr = lit.tierOf(['Systematic Review', 'Review', 'Journal Article'])
-    const rct = lit.tierOf(['Randomized Controlled Trial', 'Journal Article'])
-    const narrative = lit.tierOf(['Review', 'Journal Article'])
-    assert.ok(sr.rank < rct.rank, 'a systematic review must outrank a randomized trial')
-    assert.ok(rct.rank < narrative.rank, 'a randomized trial must outrank a narrative review')
-    assert.strictEqual(sr.label, 'Systematic review')
-    assert.strictEqual(narrative.label, 'Review')
-
-    // The guideline tier did not exist at all before Mode 3 — and the spec's one substantive
-    // sentence about this mode is "Guidelines and SRs before primary trials".
-    const guide = lit.tierOf(['Practice Guideline', 'Journal Article'])
-    assert.strictEqual(guide.rank, 1, 'a practice guideline is the top of the hierarchy')
-    assert.ok(guide.rank < sr.rank, 'guidelines come before systematic reviews')
-
-    // Still honest at the bottom: PubMed tags many real observational studies as nothing but
-    // "Journal Article", and inventing a design for them would be the exact lie we are avoiding.
-    assert.strictEqual(lit.tierOf(['Journal Article']).label, 'Other')
-
-    // THE TWO TYPES THAT OVERRIDE EVERY OTHER TYPE ON THE RECORD.
-    //
-    // PubMed ADDS "Retracted Publication" alongside the original types rather than replacing them, so
-    // a retracted trial still carries "Randomized Controlled Trial" — and first-match-wins used to
-    // find the RCT, rank it 3, and sort a withdrawn paper to the TOP of a clinical answer. Not a
-    // wrong number: a retracted one, led with as the strongest evidence available.
-    const mkTier = (pmid, types) => ({ pmid, tier: lit.tierOf(types), types })
-    const retracted = lit.tierOf(['Randomized Controlled Trial', 'Retracted Publication', 'Journal Article'])
-    assert.strictEqual(retracted.label, 'RETRACTED', 'a retracted trial is NOT an RCT, whatever else PubMed tags it')
-    assert.ok(retracted.rank > lit.tierOf(['Journal Article']).rank,
-        'a retracted paper must sort BELOW an unclassified one — it can never lead the evidence')
-    assert.strictEqual(lit.byTier([mkTier('1', ['Randomized Controlled Trial', 'Retracted Publication']), mkTier('2', ['Case Reports'])])[0].pmid, '2',
-        'a case report outranks a retracted trial')
-
-    // "Clinical Trial Protocol" startsWith "clinical trial", so it matched the rank-4 interventional
-    // tier — letting a paper that reports NO RESULTS set the evidence floor.
-    const protocol = lit.tierOf(['Clinical Trial Protocol', 'Journal Article'])
-    assert.strictEqual(protocol.label, 'Protocol', 'a protocol is not a trial: it reports no results')
-    assert.ok(protocol.rank > lit.tierOf(['Clinical Trial']).rank, 'a protocol must never outrank a completed trial')
-    console.log('tiers:        a RETRACTED trial sorts below everything; a PROTOCOL is not a clinical trial')
-
-    // ---- Pure: the sort is STABLE. ----------------------------------------------------------
-    // Within a tier, PubMed's relevance order must survive: we are re-ordering by DESIGN, not
-    // re-scoring by relevance. If this ever becomes unstable, the top of each tier silently
-    // becomes arbitrary.
-    const mk = (pmid, types) => ({ pmid, tier: lit.tierOf(types), types })
-    const sorted = lit.byTier([
-        mk('1', ['Journal Article']),
-        mk('2', ['Randomized Controlled Trial']),
-        mk('3', ['Practice Guideline']),
-        mk('4', ['Randomized Controlled Trial']),   // same tier as 2, and AFTER it in relevance
-        mk('5', ['Meta-Analysis']),
-    ])
-    assert.deepStrictEqual(
-        sorted.map(r => r.pmid), ['3', '5', '2', '4', '1'],
-        'guideline, then meta-analysis, then the two RCTs IN THEIR ORIGINAL ORDER, then the rest',
-    )
-
-    // ---- Pure: the evidence floor. ----------------------------------------------------------
-    // The sentence that changes what a clinician does next, and it is arithmetic, not inference.
-    // It must name what is ABSENT — "there is no randomized trial here" is the whole point.
-    const weak = lit.evidenceFloor([mk('1', ['Case Reports']), mk('2', ['Journal Article'])])
-    assert.ok(/case report/i.test(weak), 'the floor names the strongest design actually present')
-    assert.ok(/no randomized trial/i.test(weak), 'the floor says the RCT is missing')
-    assert.ok(/no clinical guideline/i.test(weak), 'the floor says the guideline is missing')
-    const strong = lit.evidenceFloor([mk('1', ['Practice Guideline']), mk('2', ['Randomized Controlled Trial']),
-        mk('3', ['Meta-Analysis'])])
-    assert.ok(!/no /i.test(strong), 'nothing is missing, so the floor claims nothing is missing')
-    assert.ok(/guideline/i.test(strong), strong)
-
-    // ---- Pure: PICO. ------------------------------------------------------------------------
-    // The four fields become ONE sentence, server-side. The client never composes it.
-    assert.strictEqual(
-        lit.picoQuestion({ population: 'adults with T2DM', intervention: 'SGLT2 inhibitors', comparison: 'metformin', outcome: 'CV mortality' }),
-        'In adults with T2DM, does SGLT2 inhibitors, compared with metformin, affect CV mortality?',
-    )
-    // Comparison is OPTIONAL — plenty of real clinical questions have no comparator, and requiring
-    // one invites the asker to invent it.
-    assert.strictEqual(
-        lit.picoQuestion({ population: 'adults with T2DM', intervention: 'SGLT2 inhibitors', outcome: 'CV mortality' }),
-        'In adults with T2DM, does SGLT2 inhibitors affect CV mortality?',
-    )
-    assert.ok(!lit.picoComplete({ population: 'x', intervention: 'y' }), 'Outcome is required')
-    assert.ok(lit.picoComplete({ population: 'x', intervention: 'y', outcome: 'z' }), 'Comparison is not')
-
-    // ---- LIVE: the guideline tier against a REAL PubMed record. -----------------------------
-    // The pure tests above prove the ordering. This proves the DERIVATION still lands on real
-    // data — that PubMed's guideline records really do arrive carrying a publication type we
-    // recognise, through the retrieval tool's parser, in the shape toRecord() expects. A tier
-    // table that is right about strings and wrong about records would pass everything above.
-    const guidelines = await lit.fetchRecords('"Practice Guideline"[pt] AND depression[majr]', 3, 'relevance')
-    assert.ok(guidelines.length, 'PubMed returned no practice guidelines for a query that must have them')
-    const g = guidelines[0]
-    assert.strictEqual(g.tier.rank, 1,
-        `a Practice Guideline[pt] record must land in tier 1, got "${g.design}" from types ${JSON.stringify(g.types)}`)
-    assert.ok(g.types.length, 'the RAW publication types must survive onto the record — the tier claim is only auditable if they do')
-    console.log(`guideline:    PMID ${g.pmid} -> tier ${g.tier.rank} (${g.design}) from ${JSON.stringify(g.types)}`)
-
-    // =======================================================================================
-    // THE RESULTS COLUMN — one record count per PRESS line.
-    //
-    // Asserted RELATIONALLY, never against a literal: PubMed indexes new papers daily, so pinning
-    // "line 1 == 40,132" would rot within weeks. What cannot drift is the ARITHMETIC — and the
-    // arithmetic is exactly what a librarian reads the column for.
-
-    // countRows is its own call, not part of runStrategy — the column costs 7 counts and a measured
-    // 5.5s at the retrieval tool's 500ms pacing, and it must never sit between a librarian's click
-    // and the yield they clicked for. The check exercises it the same way the route does.
-    const rowsHits = await lit.countPubmed(lit.assembleQuery(s))
-    const rc = await lit.countRows(s, rowsHits)
-    const rowsResult = { hits: rowsHits }
-    const numbered = lit.numberStrategy(s)
-    const lastRow = numbered.rows.filter(r => r.n !== null).pop()
-
-    // THE ONE THAT WOULD LIE. The final line IS the whole search, so its count IS the yield printed
-    // at the top of the panel. If these two ever disagree, one of the numbers on screen is describing
-    // a query we did not run — and there is no way for the librarian to tell which.
-    assert.strictEqual(rc[lastRow.n], rowsResult.hits,
-        'the last line of the history IS the search: its count must equal the yield')
-
-    // AND-ing can only NARROW. The combined line can never retrieve more than the smallest block it
-    // is built from — if it does, the row queries are not the queries we think they are.
-    const conceptRows = numbered.rows.filter(r => r.kind === 'combine' && /OR/.test(r.text))
-    const andRow = numbered.rows.find(r => r.kind === 'combine' && /AND/.test(r.text) && !/\[dp\]/.test(r.text))
-    for (const cr of conceptRows) {
-        assert.ok(rc[andRow.n] <= rc[cr.n],
-            `AND-ing cannot widen: line ${andRow.n} (${rc[andRow.n]}) must be <= line ${cr.n} (${rc[cr.n]})`)
-    }
-
-    // OR-ing can only WIDEN. A concept's combined line must retrieve at least as much as any single
-    // term line inside it.
-    for (const cr of conceptRows) {
-        const ci = numbered.conceptLines.findIndex(ns => ns.includes(cr.n - 1))
-        for (const termN of numbered.conceptLines[ci] || []) {
-            assert.ok(rc[cr.n] >= rc[termN],
-                `OR-ing cannot narrow: line ${cr.n} (${rc[cr.n]}) must be >= line ${termN} (${rc[termN]})`)
-        }
-    }
-
-    // An UNTICKED line has no number, so it must have no count. A count beside a line that was not
-    // searched is the same class of lie as a methods block describing an un-toggled strategy.
-    const offRows = lit.numberStrategy(untickConcept(s, 1)).rows.filter(r => r.kind === 'term' && r.n === null)
-    assert.ok(offRows.length > 0, 'unticking a concept leaves unnumbered rows')
-    console.log(`rowCounts:    line ${lastRow.n} == yield (${rowsResult.hits}); AND narrows; OR widens; unticked lines carry no count`)
-
-    // =======================================================================================
-    // THE CONCEPT / RENDERING SPLIT, and the database dialects.
-    //
-    // The whole point of the split is that a LABEL crosses a database boundary and TERMS never do.
-    // Everything below is a rule that, if it broke, would break SILENTLY — a Scopus search running
-    // without the limit the librarian asked for, or a seed check that cannot see half of Scopus.
-
-    // A limit one database cannot express is DECLARED, not dropped. Scopus indexes a document type,
-    // not a study design — probed live, DOCTYPE(rct) returns 0 — so "RCT only" has NO Scopus
-    // equivalent. If this ever starts returning terms, someone has invented a fake RCT filter and
-    // the Scopus count is quietly answering a different question from the PubMed one beside it.
-    const pmLimits = lit.buildLimits('pubmed', '5y', 'rct')
-    assert.ok(pmLimits.terms.includes('[dp]') && pmLimits.terms.includes('Randomized Controlled Trial[pt]'),
-        'PubMed expresses both limits natively')
-    assert.deepStrictEqual(pmLimits.unsupported, [], 'PubMed can express every limit the UI offers')
-
-    const scLimits = lit.buildLimits('scopus', '5y', 'rct')
-    assert.ok(scLimits.terms.includes('PUBYEAR >'), 'Scopus expresses the DATE limit natively')
-    assert.ok(!/\[dp\]|\[pt\]/.test(scLimits.terms), 'PubMed field tags must NEVER reach a Scopus query')
-    assert.deepStrictEqual(scLimits.unsupported, ['RCT only'],
-        'Scopus has no RCT document type, and MUST say so rather than run an unlimited search')
-    assert.ok(!/randomi/i.test(scLimits.terms),
-        'an inexpressible limit must not be faked with free text — that is a translation layer in disguise')
-
-    // A SEED IS NOT A PMID. A Scopus-only record (a conference paper, a non-MEDLINE journal) has no
-    // PMID at all, and those are the records Scopus is FOR — so a PMID-keyed seed check could only
-    // ever validate the half of Scopus that PubMed already covers. DOIs must parse and must render
-    // in each database's own identifier syntax.
-    const mixed = lit.parseSeeds('37314797, 10.1001/jamapsychiatry.2023.1817 not-an-id 37314797')
-    assert.deepStrictEqual(mixed, [
-        { id: '37314797', kind: 'pmid' },
-        { id: '10.1001/jamapsychiatry.2023.1817', kind: 'doi' },
-    ], 'PMIDs and DOIs both parse; junk is dropped; duplicates collapse')
-
-    // ...AND THE SCREEN MUST COUNT WITH THIS EXACT FUNCTION. The component used to keep its own
-    // PMID-only `parseSeeds` (a `/^\d{5,9}$/` filter), which meant the live "N seeds" pill and the
-    // server's validation disagreed in BOTH directions: a pasted DOI counted 0 on screen and 1 on
-    // the server, a comma-separated list of 3 counted 1, and the same PMID twice counted 2 on
-    // screen and 1 on the server. The pill is the librarian's only pre-flight evidence that their
-    // seeds registered, so a shadowing copy is a confident wrong number by construction. There is
-    // no runtime seam to assert this through (the .tsx never reaches node), so assert it at the
-    // source: the browser imports the shared parser and does not redefine one.
-    const tsx = fs.readFileSync(
-        path.join(ROOT, 'src/components/elements/Literature/LiteratureSearch.tsx'), 'utf8')
-    assert.ok(!/(?:const|function)\s+parseSeeds\b/.test(tsx),
-        'LiteratureSearch.tsx must NOT define its own parseSeeds — it shadows the shared one and the seed count stops matching what the server validates')
-    assert.ok(/^\s*parseSeeds,\s*$/m.test(tsx),
-        'LiteratureSearch.tsx must import parseSeeds from literatureSearch.strategy')
-
-    const [pmSeed, doiSeed] = mixed
-    assert.strictEqual(lit.DIALECTS.pubmed.seedQuery(pmSeed), '37314797[uid]')
-    assert.strictEqual(lit.DIALECTS.pubmed.seedQuery(doiSeed), '"10.1001/jamapsychiatry.2023.1817"[aid]')
-    assert.strictEqual(lit.DIALECTS.scopus.seedQuery(pmSeed), 'PMID(37314797)')
-    assert.strictEqual(lit.DIALECTS.scopus.seedQuery(doiSeed), 'DOI("10.1001/jamapsychiatry.2023.1817")')
-
-    // THE DOI IS QUOTED, AND ONLY A LIVE COUNT CAN PROVE IT MATTERS. A PII-style DOI carries literal
-    // parentheses, and the seed ref is only ever counted CONCATENATED (`${ref} AND (${query})`) --
-    // where those parens become PubMed's own Boolean grouping. Unquoted, a paper the search really
-    // does retrieve is reported as a MISS, and the diagnosis then blames every concept block.
-    // The control is the same paper by PMID: if that stops returning 1, this fixture rotted rather
-    // than the quoting regressing.
-    const lancet = { id: '10.1016/S0140-6736(20)30183-5', kind: 'doi' }   // PMID 31986264
-    const inQuery = q => lit.countPubmed(`${q} AND (wuhan[tiab])`)
-    assert.strictEqual(await inQuery('31986264[uid]'), 1, 'control: the query does retrieve the paper by PMID')
-    assert.strictEqual(
-        await inQuery(lit.DIALECTS.pubmed.seedQuery(lancet)), 1,
-        'a parenthesised DOI seed must survive concatenation -- unquoted it counts 0 and the paper reads as a miss',
-    )
-    console.log('dialects:     limits native per DB; Scopus DECLARES "RCT only" as inexpressible; seeds are PMID *or* DOI')
-
-    // ---- EMBASE (OVID): the database we DRAFT and never RUN. -------------------------------
-    //
-    // We have no API for it, so it produces NO NUMBER — and `strict: false` in tsconfig means the
-    // compiler will NOT stop anyone turning that null into a zero. These assertions are the only net.
-    assert.strictEqual(lit.DIALECTS.embase.countable, false, 'Embase has no API: it cannot be counted')
-    assert.strictEqual(lit.DIALECTS.pubmed.countable, true)
-    assert.strictEqual(lit.DIALECTS.scopus.countable, true)
-
-    // countIn must THROW rather than fall through to PubMed's counter. A `db === 'scopus' ? a : b`
-    // ternary would have counted an Ovid/Emtree query against PubMed and printed the answer next to
-    // the Embase strategy — a number about a different database entirely.
-    assert.throws(() => lit.countIn('embase'), /cannot be counted/i,
-        'counting an uncountable database must throw, never silently use another database s counter')
-    assert.throws(() => lit.DIALECTS.embase.seedQuery({ id: '1', kind: 'pmid' }), /cannot be counted/i,
-        'a seed check needs a count; inventing Ovid seed syntax we never verified would be worse than none')
-
-    // Ovid applies date/type limits in its OWN Limits panel, not as terms in the Boolean. So every one
-    // of them is DECLARED, and NONE is silently dropped into the query.
-    const eLimits = lit.buildLimits('embase', '5y', 'rct')
-    assert.strictEqual(eLimits.terms, '', 'no limit may be fabricated into an Ovid query')
-    assert.deepStrictEqual(eLimits.unsupported.length, 2, 'both limits are DECLARED as applied in Ovid, not dropped')
-
-    // The query is passed through VERBATIM. Ovid syntax must survive assembly untouched: mangle
-    // `exp probiotic agent/` and the line does not merely get worse, it does not run at all.
-    const ovid = {
-        db: 'embase',
-        concepts: [
-            { label: 'Probiotics', lines: [{ terms: 'exp probiotic agent/ or intestine flora/', on: true }, { terms: 'probiotic*.ti,ab,kw.', on: true }] },
-            { label: 'Depression', lines: [{ terms: 'exp depression/', on: true }] },
-        ],
-        limits: '',
-    }
-    const oq = lit.assembleQuery(ovid)
-    assert.ok(oq.includes('exp probiotic agent/') && oq.includes('probiotic*.ti,ab,kw.'),
-        'Ovid syntax must pass through assembleQuery verbatim')
-    assert.ok(!oq.includes('[tiab]') && !oq.includes('TITLE-ABS-KEY') && !/\/exp\b/.test(oq),
-        'no PubMed, Scopus or Embase.com syntax may leak into an Ovid query')
-
-    // runStrategy makes NO network call for an uncountable database, and returns null — never 0, and
-    // never an empty seed list that would read as "none of your known papers came back".
-    const eRun = await lit.runStrategy(ovid, lit.parseSeeds('37314797'), eLimits.unsupported)
-    assert.strictEqual(eRun.hits, null, 'an uncounted database yields NULL, never 0')
-    assert.deepStrictEqual(eRun.rowCounts, {}, 'no Results column without a counter')
-    assert.deepStrictEqual(eRun.seeds, [], 'a seed cannot be validated against a database we cannot count')
-
-    // The export half of this lives in the exports section below, where the renderers are loaded.
-    embaseRun = { run: eRun, unsupported: eLimits.unsupported }
-    console.log('embase(ovid): countable=false -> hits NULL not 0, no seed check, Ovid syntax passes through verbatim')
-    console.log('seed quoting: a PII-style DOI (parens and all) counts 1 inside `AND (...)`, same as its PMID')
-
-    // ---- Live Scopus. Skipped LOUDLY when the tool is not configured. ----------------------
-    if (!process.env.RECITER_SCOPUS_API_URL && !process.env.RECITER_API_BASE_URL) {
-        console.log('scopus:       SKIPPED — set RECITER_SCOPUS_API_URL to check the Scopus half')
-    } else {
-        const sc = {
-            db: 'scopus',
-            concepts: [
-                { label: 'Probiotics / microbiome', lines: [{ terms: 'TITLE-ABS-KEY(probiotic* OR synbiotic* OR "gut microbiome")', on: true }] },
-                { label: 'Depression', lines: [{ terms: 'TITLE-ABS-KEY(depress* OR "mood disorder")', on: true }] },
-            ],
-            limits: lit.buildLimits('scopus', '5y', 'any').terms,
-        }
-        const scQuery = lit.assembleQuery(sc)
-        // assembleQuery is DB-NEUTRAL: OR inside a block, AND between blocks, limits appended. If it
-        // ever grows a `switch (s.db)`, something has gone wrong upstream.
-        assert.ok(scQuery.includes('TITLE-ABS-KEY') && scQuery.includes('PUBYEAR >'),
-            'the Scopus query is native Scopus, top-level limit and all')
-
-        const scHits = await lit.countScopus(scQuery)
-        assert.ok(Number.isFinite(scHits) && scHits > 0, `expected a positive Scopus yield, got ${scHits}`)
-
-        // Narrowing must NARROW. This is the assertion that would catch a force-wrapped query: nest
-        // a top-level limit inside TITLE-ABS-KEY() and Elsevier answers 400, or worse, silently
-        // ignores it and this number does not move.
-        const scNarrower = await lit.countScopus(`${scQuery} AND DOCTYPE(ar)`)
-        assert.ok(scNarrower < scHits, 'a top-level DOCTYPE limit must actually narrow the Scopus count')
-
-        // THE SEED CHECK, IN SCOPUS. Scopus indexes the PMIDs of the records it shares with MEDLINE,
-        // which is what lets a librarian seed a Scopus strategy with the PMIDs they already have.
-        const scSeeds = await lit.validateSeeds(sc, lit.parseSeeds('37314797 99999999999'))
-        const found = scSeeds.find(x => x.id === '37314797')
-        const absent = scSeeds.find(x => x.id === '99999999999')
-        assert.ok(found && found.retrieved, 'Scopus must retrieve Nikolova 2023 for this strategy')
-
-        // NOT IN THE DATABASE IS NOT A STRATEGY BUG, and it must never be reported as one. Without
-        // this verdict a paper Scopus has never indexed "fails" every concept block at once, which
-        // reads as a catastrophically narrow query and sends a librarian off to widen a search that
-        // was never broken.
-        assert.ok(absent && absent.notInDatabase === true,
-            'a record Scopus does not index is reported as a COVERAGE gap, not as a failing block')
-        assert.ok(!absent.failingConcepts, 'a not-in-database miss must not also blame a concept block')
-        console.log(`scopus:       ${scHits} hits -> ${scNarrower} with DOCTYPE(ar); seed PMID 37314797 retrieved; a missing record reads as coverage, not as a bad query`)
-    }
-
-    // =======================================================================================
-    // EXPORTS. An export that cannot be re-run is not evidence — so the thing worth pinning is that
-    // every document carries the query, the count, the date, who ran it, and WHICH MODEL DRAFTED IT.
-    //
-    // Asserted over the BLOCKS, not over the rendered bytes. The blocks are what a document SAYS;
-    // .docx is a zip, and grepping a zip for a PubMed query proves nothing. The renderer that turns
-    // blocks into that zip is a different question and gets its own smoke test, below.
-
-    const xp = require(path.join(OUT, 'controllers/literatureExport.js'))
-
-    // Everything a document says, flattened — table cells INCLUDED, because the query, the count and
-    // the model all live in the repro TABLE, not in a paragraph. Flatten only the paragraphs and
-    // this whole section would pass while asserting nothing.
-    const said = blocks => blocks.flatMap(b =>
-        b.kind === 'table' ? [...b.head, ...b.rows.flat()] : [b.text]
-    ).join('\n')
-
-    // Prettified for the reader, but the PROFILE ID survives verbatim: the id pins the weights and
-    // agrees with the Bedrock bill, the pretty name is what a reader understands, and a journal's AI
-    // declaration wants both. An id this cannot parse must fall through UNCHANGED rather than vanish
-    // — a missing model name beats a wrong one.
-    assert.strictEqual(xp.modelLabel('us.anthropic.claude-opus-4-8'), 'Claude Opus 4.8, via AWS Bedrock')
-    assert.strictEqual(xp.modelLabel('anthropic.claude-sonnet-4-5'), 'Claude Sonnet 4.5, via AWS Bedrock')
-    assert.strictEqual(xp.modelLabel('some-model-we-have-never-seen'), 'some-model-we-have-never-seen')
-    assert.strictEqual(xp.modelLabel(''), '')
-
-    // THE FULL-LENGTH IDS, which is what BEDROCK_MODEL_ID actually holds the moment anyone pins a
-    // profile instead of taking the floating alias. Every id below is a REAL, ACTIVE profile in the
-    // account (aws bedrock list-inference-profiles). The version must stop before the release date
-    // and before the -vN revision — a regex that reads "digits and hyphens" swallows both and prints
-    // "Claude Opus 4.5.20251101." into the AI declaration of every export. There is no error worse
-    // than an invented model version, because it is indistinguishable from a real one.
-    assert.strictEqual(xp.modelLabel('us.anthropic.claude-opus-4-8-v1:0'), 'Claude Opus 4.8, via AWS Bedrock')
-    assert.strictEqual(xp.modelLabel('us.anthropic.claude-opus-4-5-20251101-v1:0'), 'Claude Opus 4.5, via AWS Bedrock')
-    assert.strictEqual(xp.modelLabel('us.anthropic.claude-sonnet-4-5-20250929-v1:0'), 'Claude Sonnet 4.5, via AWS Bedrock')
-    assert.strictEqual(xp.modelLabel('global.anthropic.claude-haiku-4-5-20251001-v1:0'), 'Claude Haiku 4.5, via AWS Bedrock')
-    assert.strictEqual(xp.modelLabel('us.anthropic.claude-opus-4-6-v1'), 'Claude Opus 4.6, via AWS Bedrock')
-    // A one-part version is a version: Sonnet 4 is not Sonnet 4.20250514.
-    assert.strictEqual(xp.modelLabel('us.anthropic.claude-sonnet-4-20250514-v1:0'), 'Claude Sonnet 4, via AWS Bedrock')
-    assert.strictEqual(xp.modelLabel('us.anthropic.claude-sonnet-5'), 'Claude Sonnet 5, via AWS Bedrock')
-
-    // THE FALLBACK, which is the whole safety property: an id whose shape we do not recognize gets
-    // NO pretty name rather than a guessed one. Legacy claude-3 ids put the generation before the
-    // family (claude-3-5-sonnet), so they do not parse — and printing the raw id is the correct,
-    // honest outcome. The id is the half that had to be right anyway.
-    assert.strictEqual(xp.modelLabel('anthropic.claude-3-5-sonnet-20240620-v1:0'), 'anthropic.claude-3-5-sonnet-20240620-v1:0')
-    assert.strictEqual(xp.modelLabel('us.anthropic.claude-3-haiku-20240307-v1:0'), 'us.anthropic.claude-3-haiku-20240307-v1:0')
-    // An unrecognized SUFFIX must also fall through whole — never a truncated guess.
-    assert.strictEqual(xp.modelLabel('us.anthropic.claude-opus-4-8-preview'), 'us.anthropic.claude-opus-4-8-preview')
-
-    // And the id itself survives verbatim in the disclosure, whatever the label does.
-    for (const id of ['us.anthropic.claude-opus-4-5-20251101-v1:0', 'anthropic.claude-3-5-sonnet-20240620-v1:0']) {
-        const decl = said(xp.strategyDoc(
-            { concepts: [{ label: 'P', lines: [{ terms: 'probiotics[tiab]', on: true }] }],
-              limits: '', query: 'probiotics[tiab]', hits: 122, runDate: '2026-07-13', seeds: [] },
-            'Q?', 'paa2013', id,
-        ))
-        assert.ok(decl.includes(id), `the profile id must survive verbatim in the declaration: ${id}`)
-        assert.ok(!/\d{8}\./.test(decl), `a release date must never be printed as a version: ${id}`)
-    }
-
-    // THE REPRODUCIBILITY INVARIANT. A synthesis pasted into a manuscript without the query behind
-    // it is an anecdote with citations — and "AI-assisted" with no model version is not a
-    // declaration. This is the assertion that says both out loud.
-    const facts = {
-        query: 'probiotics[tiab] AND depression[tiab]', hits: 122, runDate: '2026-07-13',
-        retrieved: 50, sort: 'most relevant',
-        cwid: 'paa2013', model: 'us.anthropic.claude-opus-4-8',
-    }
-    const doc = said(xp.synthesisDoc(
-        { table: [{ pmid: '37314797', study: 'Nikolova et al.', year: '2023', journal: 'JAMA Psychiatry', design: 'RCT', intervention: 'Probiotic vs placebo' }],
-          prose: 'Probiotics reduced symptoms [PMID 37314797].', floor: 'The strongest evidence retrieved is a randomized controlled trial.' },
-        facts, 'Do probiotics help depression?', { pico: true, screenedIn: 1, screenedOf: 50 },
-    ))
-    for (const must of ['probiotics[tiab] AND depression[tiab]', '122', '2026-07-13', '37314797', 'paa2013',
-                        'Claude Opus 4.8', 'us.anthropic.claude-opus-4-8']) {
-        assert.ok(doc.includes(must),
-            `the exported document must carry "${must}" — an export that cannot be re-run is not evidence`)
-    }
-
-    // THE YIELD IS NOT THE RETRIEVED COUNT, and a methods table that says it is will be copied
-    // straight into a PRISMA flow diagram. This document retrieves the top 50 of 122 and screens
-    // them; the closing line already says "1 of 50 retrieved records", so a header reading
-    // "Records retrieved: 122" contradicts its own document — and 122 is the number a co-author
-    // would write down. Both numbers, named for what they are.
-    assert.ok(doc.includes('Records identified by the query'),
-        'the yield must be labelled as the yield, never as the number of records retrieved')
-    assert.ok(/top 50 of 122, ranked by most relevant/.test(doc),
-        'a capped, ranked slice must DECLARE the cap and the ranking in the methods, not only on screen')
-    assert.ok(!/Records retrieved\D+122/.test(doc),
-        'the 122 hits must never be printed under the word "retrieved" — only 50 records were')
-    // Mode 1 counts and never retrieves, so it keeps its single row and gains no cap sentence.
-    const mode1 = said(xp.strategyDoc(
-        { concepts: [{ label: 'P', lines: [{ terms: 'probiotics[tiab]', on: true }] }], limits: '', query: 'q', hits: 122, runDate: '2026-07-13', seeds: [] },
-        'Q?', 'paa2013', 'us.anthropic.claude-opus-4-8',
-    ))
-    assert.ok(!/top \d+ of/.test(mode1), 'Mode 1 retrieves nothing, so it must not claim a retrieved slice')
-
-    // A FABRICATED CITATION MUST TRAVEL WITH THE DOCUMENT. The server already detects a PMID the
-    // model was never given; for a while it only console.error'd it and exported the clickable link
-    // anyway. The .docx is the copy that gets mailed to a co-author, so the warning belongs IN it —
-    // and above the prose, not in a footnote under it.
-    const dirty = said(xp.synthesisDoc(
-        { table: [], prose: 'Probiotics reduced symptoms [PMID 99999999].', invented: ['99999999'] },
-        facts, 'Q?', { pico: false, screenedIn: 1, screenedOf: 50 },
-    ))
-    assert.ok(/WARNING/.test(dirty) && dirty.includes('99999999'),
-        'a synthesis citing a PMID outside the evidence set must say so IN the exported document')
-    assert.ok(dirty.indexOf('WARNING') < dirty.indexOf('Probiotics reduced symptoms'),
-        'the warning must sit ABOVE the prose it contaminates — a reader must not reach the claim first')
-    assert.ok(!/WARNING/.test(doc), 'a clean synthesis must carry no warning')
-    console.log('synthesis:    an invented citation is flagged on the wire, on screen, and above the prose in the .docx')
-
-    // The strategy export must describe the TOGGLED state, never the model's draft: an unticked line
-    // was not searched, so it must not appear in a methods section that claims it was.
-    const strategyBlocks = xp.strategyDoc({
-        concepts: [
-            { label: 'Probiotics', lines: [{ terms: 'probiotics[tiab]', on: true }, { terms: 'SHOULD-NOT-APPEAR[tiab]', on: false }] },
-            { label: 'Depression', lines: [{ terms: 'depression[tiab]', on: true }] },
-        ],
-        limits: '', query: 'q', hits: 5, runDate: '2026-07-13',
-        seeds: [{ id: '1', kind: 'pmid', retrieved: false, label: 'Smith 2020', missReason: 'Depression block' }],
-    }, 'Q?', 'paa2013', 'us.anthropic.claude-opus-4-8')
-    const sDoc = said(strategyBlocks)
-    assert.ok(!sDoc.includes('SHOULD-NOT-APPEAR'), 'an UNTICKED line was never searched and must not appear in the methods')
-    assert.ok(sDoc.includes('Depression block'), 'a missed seed must carry its derived reason into the appendix')
-    // THE DISCLOSURE PEOPLE FORGET. Mode 1's QUERY is model-drafted too, not only Mode 2's prose —
-    // and this export IS the PRISMA-S appendix, so if the disclosure is not here it is nowhere.
-    assert.ok(sDoc.includes('us.anthropic.claude-opus-4-8'),
-        'the SEARCH STRATEGY export must disclose the model that drafted the query, not only the synthesis')
-
-    // THE RESULTS COLUMN, ACTUALLY PRINTED. Everything asserted about the row counts further up was
-    // asserted about the MAP; the appendix is where a PRESS reviewer reads them. Until this, no check
-    // ever handed the map to the exporter — every strategyDoc() call passed no rowCounts — so the
-    // Records column was only exercised on its BLANK branch, and a column that printed every count one
-    // line out of place would have gone green.
-    //
-    // A LINE NUMBER DOES NOT NAME A STABLE QUERY (the re-keying asserted at the top of this file), so
-    // "the right number appeared" is not the property worth checking — "each count is on ITS OWN line"
-    // is. Read the count back at the number the EXPORT derived, not at the number we counted with.
-    // The last row is the anchor: it IS the whole search, so its cell must equal the yield printed in
-    // the header above it, or the appendix and the panel are describing different queries.
-    const counted = xp.strategyDoc(
-        { ...s, query: q, hits: rowsResult.hits, runDate: '2026-07-13', seeds: [], rowCounts: rc },
-        'Do probiotics help depression?', 'paa2013', 'us.anthropic.claude-opus-4-8',
-    )
-    const recordsTable = counted.find(b => b.kind === 'table' && b.head[2] === 'Records')
-    assert.ok(recordsTable, 'the strategy appendix must carry a Records column')
-    assert.strictEqual(recordsTable.rows.length, Object.keys(rc).length,
-        'every counted line reaches the appendix — a line the librarian searched and cannot see the yield of is half an appendix')
-    for (const [n, , records] of recordsTable.rows) {
-        assert.strictEqual(records, rc[Number(n)].toLocaleString(),
-            `line ${n} must print ITS OWN count (${rc[Number(n)]}), not some other line's`)
-    }
-    assert.strictEqual(
-        recordsTable.rows[recordsTable.rows.length - 1][2], rowsResult.hits.toLocaleString(),
-        'the last line of the appendix IS the search: its count must equal the yield in the header above it',
-    )
-    console.log(`records col:  ${recordsTable.rows.length} lines printed, each carrying its own count; last line == yield (${rowsResult.hits.toLocaleString()})`)
-
-    // EMBASE (OVID): THE EXPORT IS THE ARTIFACT THAT LEAVES THE BUILDING, so it is the thing that has
-    // to say the strategy was DRAFTED and never RUN. `hits` is null here, and both of the obvious
-    // coercions are fabrications in a methods table: String(null) is the word "null", Number(null) is
-    // 0 — and "Records identified by the query: 0" beside a perfectly good strategy reads, to a reader
-    // who was not in the room, as "this search found nothing."
-    assert.ok(embaseRun, 'the Embase dialect section must have run before the exports section')
-    const eDoc = said(xp.strategyDoc(
-        { ...embaseRun.run, db: 'embase', unsupportedLimits: embaseRun.unsupported },
-        'Do probiotics help depression?', 'paa2013', 'us.anthropic.claude-opus-4-8',
-    ))
-    // `said` emits one line per table CELL, so the yield is the line straight after its label. Assert
-    // on THAT cell, not on a loose window — a regex that runs past the newline sweeps up other cells
-    // and passes (or fails) for reasons that have nothing to do with the yield.
-    const eLines = eDoc.split('\n')
-    const eYield = eLines[eLines.findIndex(l => /^Records identified by the query$/i.test(l)) + 1]
-    assert.ok(/^Not counted/i.test(eYield),
-        `the yield cell must SAY it was not counted, got: ${JSON.stringify(eYield)}`)
-    assert.ok(!/^\s*(0|null)\s*$/i.test(eYield),
-        'an uncounted database must never print 0 or "null" as its yield — that reads as "found nothing"')
-    assert.ok(/ON ITS OWN/i.test(eDoc),
-        'we cannot detect a dead Emtree heading, so the export must tell the librarian how to find one')
-    assert.ok(/Limits panel/i.test(eDoc),
-        'Ovid CAN apply these limits, in its own panel — do not tell a librarian it cannot')
-    console.log('embase export: declares DRAFTED-not-run, prints no 0/null yield, tells the librarian to test each Emtree line')
-
-    // THE RENDERER. Word is the delivery format, so a Block[] it cannot render is a download button
-    // that throws in a librarian's face. Packer really zips it — a .docx is a zip, so it opens PK.
-    const { Packer } = require('docx')
-    const { docxDoc } = require(path.join(OUT, 'controllers/literatureDocx.js'))
-    const buf = await Packer.toBuffer(docxDoc([...strategyBlocks, ...xp.synthesisDoc(
-        { table: [], prose: 'p' }, facts, 'Q?', { pico: false, screenedIn: 0, screenedOf: 0 },
-    ), { kind: 'spacer' }]))
-    assert.ok(buf.length > 1000 && buf[0] === 0x50 && buf[1] === 0x4b,
-        'a .docx is a zip and must open with PK — Word will not read anything else')
-
-    // ...AND WORD MUST SEE THE PARAGRAPHS. A .docx that OPENS is not a .docx that READS. OOXML has no
-    // line break inside <w:t>, so a raw \n there is whitespace — Word renders it as a space — and the
-    // model's 3-5 paragraph answer (bottom line / supporting evidence / what this does not establish)
-    // used to arrive as one grey slab. The screen split it; the FILE, which is the thing that leaves
-    // the building, did not. So assert on the emitted XML, not on the zip magic.
-    //
-    // jszip arrives with docx and is only used to READ the output back. If a docx bump ever drops it,
-    // skip loudly rather than fail: a missing test-only dep is not a broken export.
-    let JSZip = null
-    try { JSZip = require('jszip') } catch { /* below */ }
-    if (!JSZip) {
-        console.log('docx:         SKIPPED the paragraph check — jszip (a docx dependency) is not resolvable')
-    } else {
-        const slab = await Packer.toBuffer(docxDoc([
-            { kind: 'p', text: 'Bottom line.\n\nThe supporting evidence.\n\nWhat this does not establish.' },
-        ]))
-        const xml = await (await JSZip.loadAsync(slab)).file('word/document.xml').async('string')
-        assert.strictEqual((xml.match(/<w:p[ >]/g) || []).length, 3,
-            'a 3-paragraph synthesis must emit 3 Word paragraphs — one slab of newlines is not a document')
-        assert.ok(!/<w:t[^>]*>[^<]*\n/.test(xml),
-            'no raw newline may survive inside a <w:t>: Word reads it as a space, not a break')
-        console.log('docx:         a multi-paragraph synthesis emits real Word paragraphs, not one slab')
-    }
-
-    // The spreadsheet: a null iCite percentile is NOT a zero. Number(null) === 0, and a confident
-    // "0" against a brand-new trial is a scarlet letter we invented.
-    const sheets = xp.recordSheets(
-        [{ pmid: '1', title: 'T', authors: 'A', year: '2024', journal: 'J', design: 'RCT', tier: { rank: 3 } }],
-        {}, {}, facts,
-    )
-    const pctCol = sheets[0].head.indexOf('NIH percentile')
-    assert.strictEqual(sheets[0].rows[0][pctCol], '', 'an absent percentile exports as blank, never as 0')
-    assert.ok(sheets.some(s => s.name === 'Search'), 'the data file carries the query with it, or it cannot be accounted for later')
-    console.log('exports:      query/count/date/who/MODEL carried (strategy too); unticked lines excluded; null pct stays blank; .docx renders')
-
-    fs.rmSync(OUT, { recursive: true, force: true })
+    const { q, hits } = await mode1.run(lit)
+    await mode2.run(lit, { hits })
+    await mode3.run(lit)
+    const { rc, rowsResult } = await rowCounts.run(lit)
+    // Embase (Ovid) is asserted in two halves: the pure/dialect facts where the dialects are checked, and
+    // the EXPORT in the exports section, where the renderers are loaded. This carries the run between them.
+    const { embaseRun } = await dialects.run(lit)
+    await exportsCheck.run({ q, rc, rowsResult, embaseRun })
+
+    cleanOutDir()
     console.log('\nOK - Mode 1 (assembleQuery, numberStrategy, the empty-concept rule, the derived miss-diagnosis)'
         + '\n     Mode 2 (the design derivation, the structured-abstract join, the 50-cap, count-before-fetch,'
         + '\n             the three bands, the priced narrowings, and the escape hatch)'
         + '\n     Mode 3 (the tier order incl. the SR-below-RCT inversion, the guideline tier against a real'
         + '\n             PubMed record, the stable sort, the evidence floor, and PICO)')
 })().catch(e => {
-    fs.rmSync(OUT, { recursive: true, force: true })
+    cleanOutDir()
     console.error('\nFAILED:', e.message)
     process.exit(1)
 })
